Move data structures to new folder "models"
To have them in one place and be able to use them everywhere in the project
