+++ scraper/models/__init__.py
@@ -1,0 +1,1 @@
+from scraper.models.product import Datapoint, Info, Product, MasterProduct
