+++ scraper/scrape.py
@@ -2,9 +2,10 @@ import time
 import threading
 import logging
 from datetime import datetime
+
+from scraper.models import Info
 from scraper.domains import BaseWebsiteHandler, get_website_handler
 from scraper.filemanager import Filemanager
-from scraper.format import Info
 
 
 class Scraper:
